SentinelAI/backend/requirements.txt
# Core server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Analysis
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.11.0

# Audio analysis
librosa>=0.10.0
soundfile>=0.12.0

# Reports
reportlab>=4.0.0

# Database (optional)
pymongo>=4.6.0
motor>=3.3.0

# HTTP client
httpx>=0.25.0

# Optional heavy ML packages (install separately if needed):
# pip install torch torchvision torchaudio
# pip install tensorflow
# pip install transformers
